+++ app/helpers/recording_studio_user/profiles_helper.rb
@@ -2,6 +2,8 @@
 
 module RecordingStudioUser
   module ProfilesHelper
+    include OmniauthHelper
+
     def profile_display_name(user)
       RecordingStudioUser.display_name_for(user)
     end